+++ chapter-8/chapter-8.rb
@@ -1,0 +1,420 @@
+puts 
+puts "# = = = = = = = = = = "
+puts "# Chapter 8"
+puts "# File: chapter-8.rb"
+puts "# = = = = = = = = = = "
+puts
+
+puts 'Writing Your Own Methods'
+puts 'Part 1:'
+puts 'Hello, and thank you for taking the time to'
+puts 'help me with this experiment.  My experiment'
+puts 'has to do with the way people feel about'
+puts 'Mexican food.  Just think about Mexican food'
+puts 'and try to answer every question honestly,'
+puts 'with either a "yes" or a "no".  My experiment'
+puts 'has nothing to do with bed-wetting.'
+puts
+
+# We ask these questions, but we ignore their answers.
+
+goodAnswer = false
+while (not goodAnswer)
+  puts 'Do you like eating tacos?'
+  answer = gets.chomp.downcase
+  if (answer == 'yes' or answer == 'no')
+    goodAnswer = true
+  else
+    puts 'Please answer "yes" or "no".'
+  end
+end
+
+goodAnswer = false
+while (not goodAnswer)
+  puts 'Do you like eating burritos?'
+  answer = gets.chomp.downcase
+  if (answer == 'yes' or answer == 'no')
+    goodAnswer = true
+  else
+    puts 'Please answer "yes" or "no".'
+  end
+end
+
+# We pay attention to *this* answer, though.
+goodAnswer = false
+while (not goodAnswer)
+  puts 'Do you wet the bed?'
+  answer = gets.chomp.downcase
+  if (answer == 'yes' or answer == 'no')
+    goodAnswer = true
+    if answer == 'yes'
+      wetsBed = true
+    else
+      wetsBed = false
+    end
+  else
+    puts 'Please answer "yes" or "no".'
+  end
+end
+
+goodAnswer = false
+while (not goodAnswer)
+  puts 'Do you like eating chimichangas?'
+  answer = gets.chomp.downcase
+  if (answer == 'yes' or answer == 'no')
+    goodAnswer = true
+  else
+    puts 'Please answer "yes" or "no".'
+  end
+end
+
+puts 'Just a few more questions...'
+
+goodAnswer = false
+while (not goodAnswer)
+  puts 'Do you like eating sopapillas?'
+  answer = gets.chomp.downcase
+  if (answer == 'yes' or answer == 'no')
+    goodAnswer = true
+  else
+    puts 'Please answer "yes" or "no".'
+  end
+end
+
+# Ask lots of other questions about Mexican food.
+
+puts
+puts 'DEBRIEFING:'
+puts 'Thank you for taking the time to help with'
+puts 'this experiment.  In fact, this experiment'
+puts 'has nothing to do with Mexican food.  It is'
+puts 'an experiment about bed-wetting.  The Mexican'
+puts 'food was just there to catch you off guard'
+puts 'in the hopes that you would answer more'
+puts 'honestly. Thanks again.'
+puts
+puts wetsBed
+puts ''
+
+def sayMoo
+  puts 'mooooooo...'
+end
+
+sayMoo
+sayMoo
+puts 'coin-coin'
+sayMoo
+sayMoo
+puts ''
+
+puts 'Part 2:'
+puts 'Method Parameters'
+def sayMoo numberOfMoos
+  puts 'mooooooo...'*numberOfMoos
+end
+
+sayMoo 3
+puts 'oink-oink'
+sayMoo 2 # This should give an error because the parameter is missing.
+puts ''
+
+puts 'Part 3:'
+puts 'Local Variables'
+def doubleThis num
+  numTimes2 = num*2
+  puts num.to_s+' doubled is '+numTimes2.to_s
+end
+
+doubleThis 44
+
+def littlePest var
+  var = nil
+  puts 'HAHA! I ruined your variable!'
+end
+
+var = 'You can\'t even touch my variable!'
+littlePest var
+puts var
+
+puts ''
+
+puts 'Part 4:'
+puts 'Return Values'
+returnVal = puts 'This puts returned:'
+puts returnVal
+
+def sayMoo numberOfMoos
+  puts 'mooooooo...'*numberOfMoos
+  'yellow submarine'
+end
+
+x = sayMoo 2
+puts x
+puts ''
+
+def ask question
+  goodAnswer = false
+  while (not goodAnswer)
+    puts question
+    reply = gets.chomp.downcase
+
+    if (reply == 'yes' or reply == 'no')
+      goodAnswer = true
+      if reply == 'yes'
+        answer = true
+      else
+        answer = false
+      end
+    else
+      puts 'Please answer "yes" or "no".'
+    end
+  end
+
+  answer  # This is what we return (true or false).
+end
+
+puts 'Hello, and thank you for...'
+puts
+
+ask 'Do you like eating tacos?'      # We ignore this return value.
+ask 'Do you like eating burritos?'
+wetsBed = ask 'Do you wet the bed?'  # We save this return value.
+ask 'Do you like eating chimichangas?'
+ask 'Do you like eating sopapillas?'
+ask 'Do you like eating tamales?'
+puts 'Just a few more questions...'
+ask 'Do you like drinking horchata?'
+ask 'Do you like eating flautas?'
+
+puts
+puts 'DEBRIEFING:'
+puts 'Thank you for...'
+puts
+puts wetsBed
+
+puts ''
+
+puts 'Part 5:'
+puts 'One More Big Example'
+def englishNumber number
+  # We only want numbers from 0-100.
+  if number < 0
+    return 'Please enter a number zero or greater.'
+  end
+  if number > 100
+    return 'Please enter a number 100 or lesser.'
+  end
+
+  numString = ''  # This is the string we will return.
+
+  # "left" is how much of the number we still have left to write out.
+  # "write" is the part we are writing out right now.
+  # write and left... get it?  :)
+  left  = number
+  write = left/100          # How many hundreds left to write out?
+  left  = left - write*100  # Subtract off those hundreds.
+
+  if write > 0
+    return 'one hundred'
+  end
+
+  write = left/10          # How many tens left to write out?
+  left  = left - write*10  # Subtract off those tens.
+
+  if write > 0
+    if write == 1  # Uh-oh...
+      # Since we can't write "tenty-two" instead of "twelve",
+      # we have to make a special exception for these.
+      if    left == 0
+        numString = numString + 'ten'
+      elsif left == 1
+        numString = numString + 'eleven'
+      elsif left == 2
+        numString = numString + 'twelve'
+      elsif left == 3
+        numString = numString + 'thirteen'
+      elsif left == 4
+        numString = numString + 'fourteen'
+      elsif left == 5
+        numString = numString + 'fifteen'
+      elsif left == 6
+        numString = numString + 'sixteen'
+      elsif left == 7
+        numString = numString + 'seventeen'
+      elsif left == 8
+        numString = numString + 'eighteen'
+      elsif left == 9
+        numString = numString + 'nineteen'
+      end
+      # Since we took care of the digit in the ones place already,
+      # we have nothing left to write.
+      left = 0
+    elsif write == 2
+      numString = numString + 'twenty'
+    elsif write == 3
+      numString = numString + 'thirty'
+    elsif write == 4
+      numString = numString + 'forty'
+    elsif write == 5
+      numString = numString + 'fifty'
+    elsif write == 6
+      numString = numString + 'sixty'
+    elsif write == 7
+      numString = numString + 'seventy'
+    elsif write == 8
+      numString = numString + 'eighty'
+    elsif write == 9
+      numString = numString + 'ninety'
+    end
+
+    if left > 0
+      numString = numString + '-'
+    end
+  end
+
+  write = left  # How many ones left to write out?
+  left  = 0     # Subtract off those ones.
+
+  if write > 0
+    if    write == 1
+      numString = numString + 'one'
+    elsif write == 2
+      numString = numString + 'two'
+    elsif write == 3
+      numString = numString + 'three'
+    elsif write == 4
+      numString = numString + 'four'
+    elsif write == 5
+      numString = numString + 'five'
+    elsif write == 6
+      numString = numString + 'six'
+    elsif write == 7
+      numString = numString + 'seven'
+    elsif write == 8
+      numString = numString + 'eight'
+    elsif write == 9
+      numString = numString + 'nine'
+    end
+  end
+
+  if numString == ''
+    # The only way "numString" could be empty is if
+    # "number" is 0.
+    return 'zero'
+  end
+
+  # If we got this far, then we had a number somewhere
+  # in between 0 and 100, so we need to return "numString".
+  numString
+end
+
+puts englishNumber(  0)
+puts englishNumber(  9)
+puts englishNumber( 10)
+puts englishNumber( 11)
+puts englishNumber( 17)
+puts englishNumber( 32)
+puts englishNumber( 88)
+puts englishNumber( 99)
+puts englishNumber(100)
+puts ''
+
+def englishNumber number
+  if number < 0  # No negative numbers.
+    return 'Please enter a number that isn\'t negative.'
+  end
+  if number == 0
+    return 'zero'
+  end
+
+  # No more special cases! No more returns!
+
+  numString = ''  # This is the string we will return.
+
+  onesPlace = ['one',     'two',       'three',    'four',     'five',
+               'six',     'seven',     'eight',    'nine']
+  tensPlace = ['ten',     'twenty',    'thirty',   'forty',    'fifty',
+               'sixty',   'seventy',   'eighty',   'ninety']
+  teenagers = ['eleven',  'twelve',    'thirteen', 'fourteen', 'fifteen',
+               'sixteen', 'seventeen', 'eighteen', 'nineteen']
+
+  # "left" is how much of the number we still have left to write out.
+  # "write" is the part we are writing out right now.
+  # write and left... get it?  :)
+  left  = number
+  write = left/100          # How many hundreds left to write out?
+  left  = left - write*100  # Subtract off those hundreds.
+
+  if write > 0
+    # Now here's a really sly trick:
+    hundreds  = englishNumber write
+    numString = numString + hundreds + ' hundred'
+    # That's called "recursion". So what did I just do?
+    # I told this method to call itself, but with "write" instead of
+    # "number". Remember that "write" is (at the moment) the number of
+    # hundreds we have to write out. After we add "hundreds" to
+    # "numString", we add the string ' hundred' after it.
+    # So, for example, if we originally called englishNumber with
+    # 1999 (so "number" = 1999), then at this point "write" would
+    # be 19, and "left" would be 99. The laziest thing to do at this
+    # point is to have englishNumber write out the 'nineteen' for us,
+    # then we write out ' hundred', and then the rest of
+    # englishNumber writes out 'ninety-nine'.
+
+    if left > 0
+      # So we don't write 'two hundredfifty-one'...
+      numString = numString + ' '
+    end
+  end
+
+  write = left/10          # How many tens left to write out?
+  left  = left - write*10  # Subtract off those tens.
+
+  if write > 0
+    if ((write == 1) and (left > 0))
+      # Since we can't write "tenty-two" instead of "twelve",
+      # we have to make a special exception for these.
+      numString = numString + teenagers[left-1]
+      # The "-1" is because teenagers[3] is 'fourteen', not 'thirteen'.
+
+      # Since we took care of the digit in the ones place already,
+      # we have nothing left to write.
+      left = 0
+    else
+      numString = numString + tensPlace[write-1]
+      # The "-1" is because tensPlace[3] is 'forty', not 'thirty'.
+    end
+
+    if left > 0
+      # So we don't write 'sixtyfour'...
+      numString = numString + '-'
+    end
+  end
+
+  write = left  # How many ones left to write out?
+  left  = 0     # Subtract off those ones.
+
+  if write > 0
+    numString = numString + onesPlace[write-1]
+    # The "-1" is because onesPlace[3] is 'four', not 'three'.
+  end
+
+  # Now we just return "numString"...
+  numString
+end
+
+puts englishNumber(  0)
+puts englishNumber(  9)
+puts englishNumber( 10)
+puts englishNumber( 11)
+puts englishNumber( 17)
+puts englishNumber( 32)
+puts englishNumber( 88)
+puts englishNumber( 99)
+puts englishNumber(100)
+puts englishNumber(101)
+puts englishNumber(234)
+puts englishNumber(3211)
+puts englishNumber(999999)
+puts englishNumber(1000000000000)
+puts ''